Move ExecFlagType to MooseTypes
(refs #5680)
