+++ framework/include/base/ExecStore.h
@@ -17,7 +17,7 @@
 
 #include <vector>
 
-#include "Moose.h"
+#include "MooseTypes.h"
 #include "libmesh/libmesh_common.h"
 
 /**
